+++ cliche/commands/registry.py
@@ -5,6 +5,7 @@ from .ansi import ansi
 from .art import art
 from .ask import ask
 from .code import code
+from .create import create
 from .image import image
 from .process import kill
 from .roast import roastme
@@ -27,6 +28,7 @@ def register_commands(cli):
     cli.add_command(ask)
     cli.add_command(code)
     cli.add_command(config)
+    cli.add_command(create)
     cli.add_command(image)
     cli.add_command(kill)
     cli.add_command(models)